AJOUT TEST JUNIT TAX CATEGORY

+++ src/main/java/fr/afpa/ecommerce/bean/Tax.java
@@ -1,21 +1,22 @@
 package fr.afpa.ecommerce.bean;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class Tax extends CommonTable implements Serializable {
 
     private String name;
-    private Double value;
+    private BigDecimal value;
 
     public Tax() {
     }
 
-    public Tax(String name, Double value) {
+    public Tax(String name, BigDecimal value) {
         this.name = name;
         this.value = value;
     }
 
-    public Tax(int id, String name, Double value) {
+    public Tax(int id, String name, BigDecimal value) {
         super(id);
         this.name = name;
         this.value = value;
@@ -29,11 +30,11 @@ public class Tax extends CommonTable implements Serializable {
         this.name = name;
     }
 
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
 
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
